[WIP] Unescape string lexemes on the fly
Before this string lexemes required two allocations:

1. To store a string quickly parsed from the buffer up to the ending
  `"`.
2. To store the final string after unescaping.

This merges unescape logic inside the same loop where we look for the
ending quote, so we can use only one string as a final buffer.

Note: `self.ensure_buffer()` is the same as `self.ensure_at_least(1)`
but this way it kills almost all of the gain from the main change. No
idea why…